+++ lib/ReactViews/StandardUserInterface/MapColumn.jsx
@@ -15,9 +15,11 @@ import FeatureDetection from "terriajs-cesium/Source/Core/FeatureDetection";
 import BottomDock from "../BottomDock/BottomDock";
 import classNames from "classnames";
 import { withTranslation } from "react-i18next";
-
+import Toast from "./Toast";
+import Loader from "../Loader";
 import Styles from "./map-column.scss";
 import { observer } from "mobx-react";
+import SlideUpFadeIn from "../Transitions/SlideUpFadeIn/SlideUpFadeIn";
 
 const isIE = FeatureDetection.isInternetExplorer();
 const chromeVersion = FeatureDetection.chromeVersion();
@@ -152,6 +154,18 @@ const MapColumn = observer(
                     "map-data-count"
                   )}
                 />
+                <SlideUpFadeIn isVisible={this.props.viewState.isMapZooming}>
+                  <Toast>
+                    <Loader
+                      message={this.props.t("toast.mapIsZooming")}
+                      textProps={{
+                        style: {
+                          padding: "0 5px"
+                        }
+                      }}
+                    />
+                  </Toast>
+                </SlideUpFadeIn>
                 <div className={Styles.locationDistance}>
                   <LocationBar
                     terria={this.props.terria}